pyproject.toml
[project]
name = "multi-agents-db-dict"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.25",
    "fastmcp>=3.1.1",
    "google-adk>=1.27.3",
    "psycopg2-binary>=2.9.11",
]

[tool.pytest.ini_options]
pythonpath = ["."]
